refactor: 공통컴포넌트 폴더 위치 및 명 변경 (common->ui)
